Fix infinite loop when entering app handling a spend URI and user is not logged in.
We were incorrectly spinning in the main thread waiting for wallet.loaded but that value couldn't be updated unless we exited the thread. No we recurse back into doProcessSpendURI but enter using the global_thread first and add a sleep to give the wallet updates a chance to load data.

Fixes "Hang logging in from Foldapp link"